fix: Metrics page application is updated
