+++ frontend/src/container/MetricsApplication/ErrorRateChart.tsx
