+++ frontend/src/container/MetricsApplication/ExternalApi/ExternalApiGraph.tsx
