+++ frontend/src/container/MetricsApplication/ExternalApi/graphConfig.ts
