+++ frontend/src/container/MetricsApplication/LatencyLineChart/index.tsx
