+++ frontend/src/container/MetricsApplication/RequestRateChart.tsx
