--- frontend/src/modules/Metrics/ServiceMetrics/styles.ts
@@ -1,18 +1,0 @@
-import { Card as CardComponent, Row as RowComponent } from 'antd';
-import styled from 'styled-components';
-
-export const Card = styled(CardComponent)`
-	&&& {
-		padding: 10px;
-	}
-
-	.ant-card-body {
-		padding: 0;
-	}
-`;
-
-export const Row = styled(RowComponent)`
-	&&& {
-		padding: 1rem;
-	}
-`;
