--- frontend/src/modules/Metrics/ServiceMetricsDef.tsx
@@ -1,1 +1,0 @@
-export { ServiceMetrics as default } from './ServiceMetrics';
