--- frontend/src/modules/Metrics/TopEndpointsTable.css
@@ -1,12 +1,0 @@
-@media only screen and (min-width: 768px) {
-	.topEndpointsButton {
-		white-space: nowrap;
-		padding: 0;
-	}
-
-	.topEndpointsButton span {
-		text-overflow: ellipsis;
-		overflow: hidden;
-		max-width: 120px;
-	}
-}
