--- frontend/src/modules/Metrics/styles.ts
@@ -1,12 +1,0 @@
-import { Typography } from 'antd';
-import styled from 'styled-components';
-
-export const GraphContainer = styled.div`
-	min-height: 27vh;
-`;
-
-export const GraphTitle = styled(Typography)`
-	&&& {
-		text-align: center;
-	}
-`;
